+++ src/components/EpgProgramItem/EpgProgramItem.module.scss
@@ -1,0 +1,52 @@
+@use '../../styles/variables';
+@use '../../styles/theme';
+
+.epgProgramBox {
+  position: absolute;
+}
+
+.epgProgramContent {
+  position: relative;
+  z-index: 1;
+  display: flex;
+  justify-content: flex-start;
+  width: calc(100% - variables.$base-spacing);
+  height: calc(100% - variables.$base-spacing);
+  padding: variables.$base-spacing;
+  overflow: hidden;
+  background-color: theme.$box-bg;
+  border-radius: 4px;
+  cursor: pointer;
+  transition: all 0.4s ease-in-out;
+
+  &.isLive {
+    background-color: theme.$primary-color;
+  }
+
+  &.isActive {
+    background-color: theme.$primary-color;
+  }
+}
+
+.epgProgramTitle {
+  overflow: hidden;
+  color: theme.$title-color;
+  font-weight: 500;
+  font-size: 14px;
+  white-space: nowrap;
+  text-overflow: ellipsis;
+}
+
+.epgProgramText {
+  overflow: hidden;
+  color: theme.$text-color;
+  font-size: 12px;
+  white-space: nowrap;
+  text-overflow: ellipsis;
+}
+
+.epgProgramImage {
+  width: 100px;
+  margin-right: variables.$base-spacing;
+  border-radius: 4px;
+}
